Make storage location migration atomic
This change makes migrating the storage location
for citation data atomic by running the migration
for all items in a Zotero DB transaction. If any
problems occur during the migration, the transaction
aborts and no changes are made to the items. In
this case the storage preference also remains un-
changed and the user is alerted about the failed
migration. The alert includes a list of item titles
that caused the migration to fail, so that the user
can at least manually try to fix the broken citation
data.