Add preferences window (#176)
* Create basic preferences dialog UI

* Implement storage preference

* Make preference dialog not resizable

* Make 'note' the default storage preference

* Add a description for the storage preference

* Pass Prefs as window argument instead of waiting for them to be set from bootstrap.js

* Change preferences window wrapper .html to .xul

This makes it so that the style of the preference
window is consistent with the rest of the Zotero
application. I was not able to figure out how to
achieve this with html.

* Migrate citation storage for all items on preference change

* Fix React warning

* Add `formatString` function to preferencesDialog

* Make storage location migration atomic

This change makes migrating the storage location
for citation data atomic by running the migration
for all items in a Zotero DB transaction. If any
problems occur during the migration, the transaction
aborts and no changes are made to the items. In
this case the storage preference also remains un-
changed and the user is alerted about the failed
migration. The alert includes a list of item titles
that caused the migration to fail, so that the user
can at least manually try to fix the broken citation
data.

* Add bulletpoints in front of items that caused migration failure

* Remove obsolete comment

* Fix citation storage migration rollback issue

Co-authored-by: Dominic <dominicd7@hotmail.com>